Version 0.4.1
Reorganized GameInfo script
Redid point scoring system
Added beginnings of multiple enemy types
Adjusted LaserController to reflect changes to GameInfo
Tidied up BulletController script
Changed Enemy movement AI
Enemies harm player on collision
Added movement patterns: ZigZag and TargetedPattern

+++ Assets/Scripts/Bullet/ProjectileController.cs
@@ -4,7 +4,7 @@ using System.Collections;
 public class ProjectileController : MonoBehaviour {
 	public enum projectile_type {BULLET, LASER, LINE};
 	protected GameObject instantiationObject;
-	protected int damage;
+	protected float damage;
 
 	// Use this for initialization
 	void Start () {
